+++ src/interfaces/ast/map-ast.ts
@@ -26,7 +26,7 @@ export type MapNodeKind =
 
 export interface MapASTNodeBase {
   kind: MapNodeKind;
-  location?: LocationSpan | undefined;
+  location?: LocationSpan;
 }
 
 // ATOMS
@@ -47,11 +47,14 @@ export interface ObjectLiteralNode extends MapASTNodeBase {
   fields: AssignmentNode[];
 }
 
+/**
+ * Jessie expression node: `<expression>`
+ */
 export interface JessieExpressionNode extends MapASTNodeBase {
   kind: 'JessieExpression';
   expression: string;
-  source?: string | undefined;
-  sourceMap?: string | undefined;
+  source?: string;
+  sourceMap?: string;
 }
 
 /**
@@ -59,8 +62,8 @@ export interface JessieExpressionNode extends MapASTNodeBase {
  */
 export interface InlineCallNode extends MapASTNodeBase {
   kind: 'InlineCall';
-  condition?: ConditionAtomNode | undefined;
-  iteration?: IterationAtomNode | undefined;
+  condition?: ConditionAtomNode;
+  iteration?: IterationAtomNode;
   operationName: string;
   arguments: AssignmentNode[];
 }
@@ -106,7 +109,7 @@ export interface IterationAtomNode extends MapASTNodeBase {
  */
 export interface OutcomeStatementNode extends MapASTNodeBase {
   kind: 'OutcomeStatement';
-  condition?: ConditionAtomNode | undefined;
+  condition?: ConditionAtomNode;
   isError: boolean;
   terminateFlow: boolean;
   value: LiteralNode;
@@ -119,7 +122,7 @@ export interface OutcomeStatementNode extends MapASTNodeBase {
  */
 export interface SetStatementNode extends MapASTNodeBase {
   kind: 'SetStatement';
-  condition?: ConditionAtomNode | undefined;
+  condition?: ConditionAtomNode;
   assignments: AssignmentNode[];
 }
 
@@ -128,30 +131,33 @@ export interface SetStatementNode extends MapASTNodeBase {
  */
 export interface CallStatementNode extends MapASTNodeBase {
   kind: 'CallStatement';
-  iteration?: IterationAtomNode | undefined;
-  condition?: ConditionAtomNode | undefined;
+  iteration?: IterationAtomNode;
+  condition?: ConditionAtomNode;
   operationName: string;
   arguments: AssignmentNode[];
   statements: (SetStatementNode | OutcomeStatementNode)[];
 }
 
+/**
+ * Http security requirement: `security <id>`
+ */
 export type HttpSecurityRequirement = {
   id: string;
   /** Optional scheme information. */
-  scheme?: 'apikey' | 'basic' | 'bearer' | undefined;
+  scheme?: 'apikey' | 'basic' | 'bearer';
 };
 
 /**
  * Request definition for http:
- * `request <?contentType> <?contentLanguage> { <?query> <?headers? <?body> }`
+ * `request <?contentType> <?contentLanguage> { <?query> <?headers> <?body> }`
  */
 export interface HttpRequestNode extends MapASTNodeBase {
   kind: 'HttpRequest';
-  contentType?: string | undefined;
-  contentLanguage?: string | undefined;
-  query?: ObjectLiteralNode | undefined;
-  headers?: ObjectLiteralNode | undefined;
-  body?: LiteralNode | undefined;
+  contentType?: string;
+  contentLanguage?: string;
+  query?: ObjectLiteralNode;
+  headers?: ObjectLiteralNode;
+  body?: LiteralNode;
   security: HttpSecurityRequirement[];
 }
 
@@ -165,21 +171,21 @@ export interface HttpResponseHandlerNode extends MapASTNodeBase {
    * @TJS-minimum 200
    * @TJS-maximum 599
    **/
-  statusCode?: number | undefined;
-  contentType?: string | undefined;
-  contentLanguage?: string | undefined;
+  statusCode?: number;
+  contentType?: string;
+  contentLanguage?: string;
   statements: (SetStatementNode | OutcomeStatementNode)[];
 }
 
+/**
+ * Http call statement: `http <method> <?service> <url> { <?request> <...responseHandlers> }`
+ */
 export interface HttpCallStatementNode extends MapASTNodeBase {
   kind: 'HttpCallStatement';
   method: string;
-  serviceId?: string | undefined;
-  /**
-   * @format uri-reference
-   **/
+  serviceId?: string;
   url: string;
-  request?: HttpRequestNode | undefined;
+  request?: HttpRequestNode;
   responseHandlers: HttpResponseHandlerNode[];
 }
 
@@ -191,6 +197,9 @@ export type Substatement =
   | CallStatementNode
   | HttpCallStatementNode;
 
+/**
+ * Map definition: `map <name> { <...statements> }`
+ */
 export interface MapDefinitionNode extends MapASTNodeBase, DocumentedNode {
   kind: 'MapDefinition';
 
@@ -205,6 +214,9 @@ export interface MapDefinitionNode extends MapASTNodeBase, DocumentedNode {
   statements: Substatement[];
 }
 
+/**
+ * Operation definition: `operation <name> { <...statements> }`
+ */
 export interface OperationDefinitionNode
   extends MapASTNodeBase,
     DocumentedNode {
@@ -218,13 +230,18 @@ export interface OperationDefinitionNode
 
 // DOCUMENT
 
+/**
+ * Map header:
+ * `profile = <profile>`
+ * `provider = <provider>`
+ */
 export interface MapHeaderNode extends MapASTNodeBase, DocumentedNode {
   kind: 'MapHeader';
   profile: {
     /**
      * @pattern require('./utils').DOCUMENT_NAME_RE_SOURCE
      **/
-    scope?: string | undefined;
+    scope?: string;
     /**
      * @pattern require('./utils').DOCUMENT_NAME_RE_SOURCE
      **/
@@ -244,11 +261,11 @@ export interface MapHeaderNode extends MapASTNodeBase, DocumentedNode {
        * @TJS-minimum 0
        * @TJS-type integer
        **/
-      patch?: number | undefined;
+      patch?: number;
       /**
        * @pattern require('./utils').DOCUMENT_NAME_RE_SOURCE
        **/
-      label?: string | undefined;
+      label?: string;
     };
   };
 
@@ -259,10 +276,16 @@ export interface MapHeaderNode extends MapASTNodeBase, DocumentedNode {
   /**
    * @pattern require('./utils').DOCUMENT_NAME_RE_SOURCE
    **/
-  variant?: string | undefined;
+  variant?: string;
 }
 
+/**
+ * Map document:
+ * `<header>`
+ * `<...definitions>`
+ */
 export interface MapDocumentNode extends MapASTNodeBase {
+  /** @$ref AstMetadata */
   astMetadata: AstMetadata;
   kind: 'MapDocument';
   header: MapHeaderNode;