+++ src/interfaces/ast/source.ts
@@ -22,30 +22,29 @@ export type LocationSpan = {
 
 /** Node preceded by documenting string literal */
 export interface DocumentedNode {
-  documentation?:
-    | {
-        title: string;
-        description?: string | undefined;
-        location?: LocationSpan | undefined;
-      }
-    | undefined;
+  documentation?: {
+    title: string;
+    description?: string;
+    location?: LocationSpan;
+  };
 }
 
 /**
  * Information about AST and Parser used to compile provided AST
+ * @$id AstMetadata
  */
 export interface AstMetadata {
   astVersion: {
     major: number;
     minor: number;
     patch: number;
-    label?: string | undefined;
+    label?: string;
   };
   parserVersion: {
     major: number;
     minor: number;
     patch: number;
-    label?: string | undefined;
+    label?: string;
   };
   sourceChecksum: string;
 }