+++ src/interfaces/providerjson/providerjson.utils.test.ts
@@ -1,4 +1,3 @@
-import { AssertionError } from '../../error';
 import { IntegrationParameter, prepareProviderParameters } from '.';
 import {
   ApiKeyPlacement,
@@ -193,12 +192,7 @@ describe('ProviderJsonDocument', () => {
       }`;
     expect(() => {
       assertProviderJson(JSON.parse(providerJson));
-    }).toThrowError(
-      new AssertionError(
-        `Provider JSON validation failed at input: expected 'name' in object, found: {\n  services: [ { baseUrl: 'https://swapi.dev/api', id: 'swapidev' } ],\n  defaultService: 'swapidev'\n}`,
-        []
-      )
-    );
+    }).toThrow('$: must have required property "name"');
   });
 
   it('throws error on document with invalid name', () => {
@@ -214,7 +208,7 @@ describe('ProviderJsonDocument', () => {
       }`;
     expect(() => {
       assertProviderJson(JSON.parse(providerJson));
-    }).toThrowError(new AssertionError(`invalid provider name`, []));
+    }).toThrow('$.name: must match pattern');
   });
 
   it('throws error on document with missing services', () => {
@@ -224,12 +218,7 @@ describe('ProviderJsonDocument', () => {
       }`;
     expect(() => {
       assertProviderJson(JSON.parse(providerJson));
-    }).toThrowError(
-      new AssertionError(
-        `Provider JSON validation failed at input: expected 'services' in object, found: { name: 'swapidev', defaultService: 'swapidev' }`,
-        []
-      )
-    );
+    }).toThrow('$: must have required property "services"');
   });
 
   it('throws error on document with missing property in services', () => {
@@ -244,12 +233,7 @@ describe('ProviderJsonDocument', () => {
       }`;
     expect(() => {
       assertProviderJson(JSON.parse(providerJson));
-    }).toThrowError(
-      new AssertionError(
-        `Provider JSON validation failed at input.services.[0]: expected 'baseUrl' in object, found: { id: 'swapidev' }`,
-        []
-      )
-    );
+    }).toThrow('$.services.0: must have required property "baseUrl"');
   });
 
   it('throws error on document with missing defaultService', () => {
@@ -264,12 +248,7 @@ describe('ProviderJsonDocument', () => {
       }`;
     expect(() => {
       assertProviderJson(JSON.parse(providerJson));
-    }).toThrowError(
-      new AssertionError(
-        `Provider JSON validation failed at input: expected 'defaultService' in object, found: {\n  name: 'swapidev',\n  services: [ { baseUrl: 'https://swapi.dev/api', id: 'swapidev' } ]\n}`,
-        []
-      )
-    );
+    }).toThrow('$: must have required property "defaultService"');
   });
 
   it('throws error on document with missing id property in securitySchemes', () => {