+++ src/interfaces/superjson/superjson.ts
@@ -21,6 +21,7 @@ export type UriPath = string;
 // Retry policy
 export enum OnFail {
   NONE = 'none',
+  SIMPLE = 'simple',
   CIRCUIT_BREAKER = 'circuit-breaker',
 }
 
@@ -28,76 +29,114 @@ export enum BackoffKind {
   EXPONENTIAL = 'exponential',
 }
 
+export type BackoffPolicy =
+  | BackoffKind.EXPONENTIAL
+  | {
+      kind: BackoffKind.EXPONENTIAL;
+      /**
+       * @TJS-minimum 0
+       * @TJS-type integer
+       **/
+      start?: number;
+      /**
+       * @TJS-minimum 0
+       * @TJS-type integer
+       **/
+      factor?: number;
+    };
+
 /**
- * RetryPolicy per usecase values.
+ * Retry policy configuration.
  */
 export type RetryPolicy =
   | OnFail.NONE
-  | OnFail.CIRCUIT_BREAKER
   | {
       kind: OnFail.NONE;
     }
+  | OnFail.SIMPLE
+  | {
+      kind: OnFail.SIMPLE;
+      /**
+       * @TJS-minimum 0
+       * @TJS-type integer
+       **/
+      maxContiguousRetries?: number;
+      /**
+       * @TJS-minimum 0
+       * @TJS-type integer
+       **/
+      requestTimeout?: number;
+    }
+  | OnFail.CIRCUIT_BREAKER
   | {
       kind: OnFail.CIRCUIT_BREAKER;
       /**
        * @TJS-minimum 0
        * @TJS-type integer
        **/
-      maxContiguousRetries?: number | undefined;
+      maxContiguousRetries?: number;
       /**
        * @TJS-minimum 0
        * @TJS-type integer
        **/
-      requestTimeout?: number | undefined;
-      backoff?:
-        | BackoffKind.EXPONENTIAL
-        | {
-            kind: BackoffKind.EXPONENTIAL;
-            /**
-             * @TJS-minimum 0
-             * @TJS-type integer
-             **/
-            start?: number | undefined;
-            /**
-             * @TJS-minimum 0
-             * @TJS-type integer
-             **/
-            factor?: number | undefined;
-          }
-        | undefined;
+      openTime?: number;
+      /**
+       * @TJS-minimum 0
+       * @TJS-type integer
+       **/
+      requestTimeout?: number;
+      backoff?: BackoffPolicy;
     };
+
+export type NormalizedBackoffPolicy = {
+  kind: BackoffKind.EXPONENTIAL;
+  /**
+   * @TJS-minimum 0
+   * @TJS-type integer
+   **/
+  start?: number;
+  /**
+   * @TJS-minimum 0
+   * @TJS-type integer
+   **/
+  factor?: number;
+};
 
 export type NormalizedRetryPolicy =
   | {
       kind: OnFail.NONE;
     }
   | {
+      kind: OnFail.SIMPLE;
+      /**
+       * @TJS-minimum 0
+       * @TJS-type integer
+       **/
+      maxContiguousRetries?: number;
+      /**
+       * @TJS-minimum 0
+       * @TJS-type integer
+       **/
+      requestTimeout?: number;
+    }
+  | {
       kind: OnFail.CIRCUIT_BREAKER;
       /**
        * @TJS-minimum 0
        * @TJS-type integer
        **/
-      maxContiguousRetries?: number | undefined;
+      maxContiguousRetries?: number;
       /**
        * @TJS-minimum 0
        * @TJS-type integer
        **/
-      requestTimeout?: number | undefined;
-      backoff?:
-        | {
-            kind: BackoffKind.EXPONENTIAL;
-            /**
-             * @TJS-minimum 0
-             * @TJS-type integer
-             **/
-            start?: number | undefined;
-            /**
-             * @TJS-minimum 0
-             * @TJS-type integer
-             **/
-            factor?: number | undefined;
-          }
-        | undefined;
+      openTime?: number;
+      /**
+       * @TJS-minimum 0
+       * @TJS-type integer
+       **/
+      requestTimeout?: number;
+      backoff: NormalizedBackoffPolicy;
     };
 
 /**
@@ -105,8 +144,8 @@ export type NormalizedRetryPolicy =
  */
 export type UsecaseDefaults = {
   [usecase: string]: {
-    input?: { [key: string]: unknown } | undefined;
-    providerFailover?: boolean | undefined;
+    input?: { [key: string]: unknown };
+    providerFailover?: boolean;
   };
 };
 
@@ -122,8 +161,8 @@ export type NormalizedUsecaseDefaults = {
  */
 export type ProfileProviderDefaults = {
   [provider: string]: {
-    input?: { [key: string]: unknown } | undefined;
-    retryPolicy?: RetryPolicy | undefined;
+    input?: { [key: string]: unknown };
+    retryPolicy?: RetryPolicy;
   };
 };
 
@@ -138,14 +177,14 @@ export type NormalizedProfileProviderDefaults = {
  * Provider settings for specific profile.
  */
 export type ProfileProviderSettings = {
-  defaults?: ProfileProviderDefaults | undefined;
+  defaults?: ProfileProviderDefaults;
 } & (
   | {
       file: string;
     }
   | {
-      mapVariant?: string | undefined;
-      mapRevision?: string | undefined;
+      mapVariant?: string;
+      mapRevision?: string;
     }
 );
 
@@ -155,8 +194,8 @@ export type NormalizedProfileProviderSettings =
       defaults: NormalizedProfileProviderDefaults;
     }
   | {
-      mapVariant?: string | undefined;
-      mapRevision?: string | undefined;
+      mapVariant?: string;
+      mapRevision?: string;
       defaults: NormalizedProfileProviderDefaults;
     };
 
@@ -169,11 +208,9 @@ export type ProfileProviderEntry = UriPath | ProfileProviderSettings;
  * Expanded profile settings for one profile id.
  */
 export type ProfileSettings = {
-  priority?: string[] | undefined;
-  defaults?: UsecaseDefaults | undefined;
-  providers?:
-    | { [provider: string]: UriPath | ProfileProviderEntry }
-    | undefined;
+  priority?: string[];
+  defaults?: UsecaseDefaults;
+  providers?: { [provider: string]: UriPath | ProfileProviderEntry };
 } & (
   | {
       version: SemanticVersion;
@@ -208,19 +245,32 @@ export type IdBase = {
   id: string;
 };
 
+/**
+ * @$id ApiKeySecurityValues
+ **/
 export type ApiKeySecurityValues = IdBase & {
   apikey: string;
 };
 
+/**
+ * @$id BasicAuthSecurityValues
+ **/
 export type BasicAuthSecurityValues = IdBase & {
   username: string;
   password: string;
 };
 
+/**
+ * @$id BearerTokenSecurityValues
+ **/
 export type BearerTokenSecurityValues = IdBase & {
   token: string;
 };
 
+/**
+ * Security values for digest security scheme
+ * @$id DigestSecurityValues
+ **/
 export type DigestSecurityValues = IdBase & {
   username: string;
   password: string;
@@ -239,13 +289,13 @@ export type SecurityValues =
  * Expanded provider settings for one provider name.
  */
 export type ProviderSettings = {
-  file?: string | undefined;
-  security?: SecurityValues[] | undefined;
-  parameters?: { [key: string]: string } | undefined;
+  file?: string;
+  security?: SecurityValues[];
+  parameters?: { [key: string]: string };
 };
 
 export type NormalizedProviderSettings = {
-  file?: string | undefined;
+  file?: string;
   security: SecurityValues[];
   parameters: { [key: string]: string };
 };
@@ -253,8 +303,8 @@ export type NormalizedProviderSettings = {
 export type ProviderEntry = UriPath | ProviderSettings;
 
 export type SuperJsonDocument = {
-  profiles?: { [profile: string]: ProfileEntry } | undefined;
-  providers?: { [provider: string]: ProviderEntry } | undefined;
+  profiles?: { [profile: string]: ProfileEntry };
+  providers?: { [provider: string]: ProviderEntry };
 };
 
 export type NormalizedSuperJsonDocument = {
@@ -269,7 +319,7 @@ export type AnonymizedSuperJsonDocument = {
       version: SemanticVersion | 'file';
       providers: {
         provider: string;
-        priority?: number | undefined;
+        priority?: number;
         version: SemanticVersion | 'file';
       }[];
     }